+++ webServices/routerRespHandler.js
@@ -2,7 +2,7 @@ const fs = require('fs');
 
 const index = (response) => {
   response.writeHead(200, { 'Content-Type': 'text/html' });
-  const myReadStrm = fs.createReadStream(__dirname + '/index.html', 'utf8');
+  const myReadStrm = fs.createReadStream(__dirname + '/responseView/index.html', 'utf8');
   myReadStrm.pipe(response);
 };
